test(MSW): move get-user callback into its own file to have a clean handlers array
