+++ src/mocks/handlers.ts
@@ -1,67 +1,6 @@
-import {http, HttpResponse} from "msw";
-import {User} from "../types/types";
-import {createLeadUser} from "./createUsers";
-
-interface GetUserRequestBody {
-  userId: number;
-  username: string;
-}
+import {http} from "msw";
+import {getUserResolver} from "./getUserResolver";
 
 const backendUrl = import.meta.env.VITE_BACKEND_URL;
 
-const getLeadUser = () => {
-  const numChats = parseInt(import.meta.env.VITE_NUM_CHATS || "1", 10);
-  return createLeadUser(numChats);
-};
-
-const newUser: Partial<User> = {
-  id: 2,
-  name: "New User",
-  chats: [],
-  has_profile: false,
-};
-
-const inviteeUser: Partial<User> = {
-  id: 3,
-  name: "Invitee User",
-  chats: [
-    {
-      lead_id: 2,
-      agreed_users: [1],
-      name: "Invitee Chat",
-      id: 2,
-      status: "pending",
-      words: 50,
-      users: [],
-    },
-  ],
-  has_profile: false,
-};
-
-export const handlers = [
-  http.post(`${backendUrl}/get-user`, async ({request}) => {
-    const json = await request.json();
-
-    if (!json || typeof json !== "object") {
-      return new HttpResponse("Invalid request body", {
-        status: 400,
-        headers: {"Content-Type": "application/json"},
-      });
-    }
-    const body = json as GetUserRequestBody;
-    const {userId} = body;
-    switch (userId) {
-      case 1:
-        return HttpResponse.json(getLeadUser());
-      case 2:
-        return HttpResponse.json(newUser);
-      case 3:
-        return HttpResponse.json(inviteeUser);
-      default:
-        return new HttpResponse("User not found", {
-          status: 404,
-          headers: {"Content-Type": "application/json"},
-        });
-    }
-  }),
-];
+export const handlers = [http.post(`${backendUrl}/get-user`, getUserResolver)];
